Small tweak to Snow Leopard makefile and update Ubuntu version to (almost) match; includes PDmenu workaround.
Snow Leopard makefile includes MacPorts X11 headers, instead of system.
Removed duplicated libpng(12).a.
Static compiling of X11 does not work on Ubuntu (due to a 'known problem with modular X11').
Otherwise as close as possible to Snow Leopard build.
A bug in the linked libraries(?) caused a segmentation fault if the user did not specifically select 'none' in the pull down menu.